LayoutTests/imported/w3c/web-platform-tests/service-workers/service-worker/claim-worker-fetch.https-expected.txt


PASS fetch() in Worker should be intercepted after the client is claimed. 
